mon-projet/ElectronBackendDir/_internal/src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
websockets==12.0
python-multipart==0.0.6
pydantic==2.4.2
psutil==5.9.6
asyncio-mqtt==0.16.1
pydantic-settings==2.0.3
motor==3.3.2
pymongo==4.6.2
PyYAML==6.0.1
pytest-asyncio==0.23.6
fpdf2==2.7.9
